+++ lib/scalingo/billing.rb
@@ -2,5 +2,10 @@ require "scalingo/api/client"
 
 module Scalingo
   class Billing < API::Client
+    require "scalingo/billing/profile"
+
+    register_handlers!(
+      profile: Profile,
+    )
   end
 end